@@ -1,11 +1,6 @@
-import 'package:shared_preferences/shared_preferences.dart';
-
 import 'boolean_parameter.dart';
 
 class FirstRun extends BooleanParameter {
-  FirstRun({required SharedPreferences sharedPreferences})
-      : super(
-            sharedPreferences: sharedPreferences,
-            key: 'options.firstRun',
-            defaultValue: false);
+  FirstRun({required super.sharedPreferences})
+      : super(key: 'options.firstRun', defaultValue: false);
 }
